requirements.txt
streamlit
azure-ai-formrecognizer
openai
python-dotenv
fastapi
uvicorn azure-cognitiveservices-speech
python-multipart
azure-cognitiveservices-speech
